refacto and add parseTranslate
