+++ examples/index.ts
@@ -1,4 +1,35 @@
 import lorrainjs from "../src/index";
+
+const toParseLog = [
+  `Toto 2`,
+  '{Toto 3 : "Salut"}  ',
+  '  {Toto35 : "Salut"}',
+  "  {Toto4 : 'Salut'}",
+  "  {'Toto5' : 'Salut'}",
+  '  {"Toto 6" : "Salut"} ',
+  '  {"Toto 7" : "Salut" ',
+  '  {"Toto 8" : "Salut} ',
+  `  {'Toto9' : "Salut"}  `,
+  `  ['Toto95', "Salut"]  `,
+  '  {{g: 9} : "Salut} ',
+  '  { test: "Bonjour", ignore: 456}',
+  `  { test: 'Bonjour', "ignore": "456"}`,
+  ` { test: 'Bonjour', "complex": {a: 2, b: "banane", c: 'hey'}}   `,
+  `{
+    a: "Salut Jacky, tu veux une noix de cajou ?",
+    c: 2.09,
+    d: "Le cheval de Sylvie.",
+  }`,
+  "124  ",
+  "125",
+];
+toParseLog.map((item) => {
+  console.log(item);
+  const result = lorrainjs.parseTranslate(item);
+  console.log(result);
+  console.log(typeof result);
+  console.log("------------------------------");
+});
 console.log(`Toto`);
 lorrainjs.initLog();
 lorrainjs.initError();
